Added a settings menu for audio
