--- new-game-project/settingsMenu/new-baby/.gitattributes
@@ -1,2 +1,0 @@
-# Normalize EOL for all files that Git considers text files.
-* text=auto eol=lf
