--- new-game-project/settingsMenu/new-baby/.gitignore
@@ -1,3 +1,0 @@
-# Godot 4+ specific ignores
-.godot/
-/android/
