--- new-game-project/settingsMenu/new-baby/New Recording 6.wav.import
@@ -1,24 +1,0 @@
-[remap]
-
-importer="wav"
-type="AudioStreamWAV"
-uid="uid://hj30n7fl3k3c"
-path="res://.godot/imported/New Recording 6.wav-8bdb0906358fbce780af7a1ee105f373.sample"
-
-[deps]
-
-source_file="res://New Recording 6.wav"
-dest_files=["res://.godot/imported/New Recording 6.wav-8bdb0906358fbce780af7a1ee105f373.sample"]
-
-[params]
-
-force/8_bit=false
-force/mono=false
-force/max_rate=false
-force/max_rate_hz=44100
-edit/trim=false
-edit/normalize=false
-edit/loop_mode=0
-edit/loop_begin=0
-edit/loop_end=-1
-compress/mode=0
