--- new-game-project/settingsMenu/new-baby/New baby.sln
@@ -1,19 +1,0 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "New baby", "New baby.csproj", "{E4734B24-B639-48AF-AE43-60D169E5C9E3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-	Debug|Any CPU = Debug|Any CPU
-	ExportDebug|Any CPU = ExportDebug|Any CPU
-	ExportRelease|Any CPU = ExportRelease|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.ExportDebug|Any CPU.ActiveCfg = ExportDebug|Any CPU
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
-		{E4734B24-B639-48AF-AE43-60D169E5C9E3}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
-	EndGlobalSection
-EndGlobal
