--- new-game-project/settingsMenu/new-baby/project.godot
@@ -1,20 +1,0 @@
-; Engine configuration file.
-; It's best edited using the editor UI and not directly,
-; since the parameters that go here are not all obvious.
-;
-; Format:
-;   [section] ; section goes between []
-;   param=value ; assign values to parameters
-
-config_version=5
-
-[application]
-
-config/name="New baby"
-run/main_scene="res://settings_menu.tscn"
-config/features=PackedStringArray("4.3", "C#", "Forward Plus")
-config/icon="res://icon.svg"
-
-[dotnet]
-
-project/assembly_name="New baby"
