--- new-game-project/settingsMenu/new-baby/settings_menu.tscn
@@ -1,48 +1,0 @@
-[gd_scene load_steps=3 format=3 uid="uid://ckhrl0wcwu07u"]
-
-[ext_resource type="Script" path="res://SettingsMenu.cs" id="1_kadyo"]
-[ext_resource type="Script" path="res://VolumeControl.cs" id="2_he4dm"]
-
-[node name="SettingsMenu" type="Control"]
-layout_mode = 3
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-script = ExtResource("1_kadyo")
-
-[node name="MarginContainer" type="MarginContainer" parent="."]
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer"]
-layout_mode = 2
-
-[node name="Label" type="Label" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Volume"
-
-[node name="Volume Control" type="HSlider" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-script = ExtResource("2_he4dm")
-
-[node name="Label2" type="Label" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Mute"
-
-[node name="Mute" type="CheckBox" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-
-[node name="ExitSettings" type="Button" parent="MarginContainer/VBoxContainer"]
-layout_mode = 2
-text = "Exit Settings Menu
-"
-
-[connection signal="value_changed" from="MarginContainer/VBoxContainer/Volume Control" to="." method="_on_h_slider_value_changed"]
-[connection signal="toggled" from="MarginContainer/VBoxContainer/Mute" to="." method="_on_mute_toggled"]
-[connection signal="pressed" from="MarginContainer/VBoxContainer/ExitSettings" to="." method="_on_exit_settings_pressed"]
