--- new-game-project/settingsMenu/new-baby/sprite_movement.tscn
@@ -1,26 +1,0 @@
-[gd_scene load_steps=4 format=3 uid="uid://w1o3fhxsmgoe"]
-
-[ext_resource type="Script" path="res://SpriteMovement.cs" id="1_od2o6"]
-[ext_resource type="Texture2D" uid="uid://v4qfe38wq8kn" path="res://icon.svg" id="1_w7kiu"]
-[ext_resource type="AudioStream" uid="uid://hj30n7fl3k3c" path="res://New Recording 6.wav" id="3_42o8f"]
-
-[node name="Sprite movement" type="Node2D"]
-script = ExtResource("1_od2o6")
-
-[node name="Icon" type="Sprite2D" parent="."]
-position = Vector2(434, 239)
-texture = ExtResource("1_w7kiu")
-
-[node name="goToSettings" type="Button" parent="."]
-offset_left = 989.0
-offset_top = 22.0
-offset_right = 1126.0
-offset_bottom = 84.0
-text = "Settings"
-
-[node name="NewRecording6" type="AudioStreamPlayer" parent="."]
-stream = ExtResource("3_42o8f")
-volume_db = -10.0
-autoplay = true
-
-[connection signal="pressed" from="goToSettings" to="." method="_on_button_pressed"]
